+++ frontend/src/app/app-routing.module.ts
@@ -1,12 +1,18 @@
+import { AuthGuard } from './guard/auth.guard';
+import { GuestGuard } from './guard/guest.guard';
+import { LoginComponent } from './login/login.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { DiarioErrosComponent } from './components/diario-erros/diario-erros.component';
 import { LoginSuccessComponent } from '@nuvem/angular-base';
 
 const routes: Routes = [
+    { path: '', redirectTo: 'login', pathMatch: 'full' },
+    { path: 'admin', loadChildren: () => import('./admin/admin.module')
+    .then(m => m.AdminModule), canActivate: [AuthGuard] },
+    { path: 'login', component: LoginComponent, canActivate: [GuestGuard]},
     { path: 'diario-erros', component: DiarioErrosComponent, data: { breadcrumb: 'Diário de Erros'} },
-    { path: 'login-success', component: LoginSuccessComponent },
-    { path: 'usuarios', loadChildren: () => import('./usuario/usuario.module').then(m => m.UsuarioModule)}
+    { path: 'login-success', component: LoginSuccessComponent}
 ];
 
 @NgModule({
