+++ frontend/src/app/app.module.ts
@@ -1,6 +1,7 @@
+import { TokenInterceptor } from './interceptors/token.interceptor';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NgModule } from '@angular/core';
+import { LOCALE_ID, NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
@@ -8,23 +9,25 @@ import { AppTopbarComponent } from './components/topbar/app.topbar.component';
 import { AppFooterComponent } from './components/footer/app.footer.component';
 import { LocationStrategy, HashLocationStrategy } from '@angular/common';
 import { environment } from '../environments/environment';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { PageNotificationModule, BreadcrumbModule, MenuModule, ErrorStackModule } from '@nuvem/primeng-components';
 import { ErrorModule, SecurityModule, VersionTagModule } from '@nuvem/angular-base';
 import { DiarioErrosComponent } from './components/diario-erros/diario-erros.component';
 import { BlockUIModule } from 'ng-block-ui';
+import { LoginComponent } from './login/login.component';
+import { ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
     declarations: [
         AppComponent,
-        AppTopbarComponent,
-        AppFooterComponent,
-        DiarioErrosComponent
+        DiarioErrosComponent,
+        LoginComponent
     ],
     imports: [
         BlockUIModule.forRoot({
             message: "Carregando..."
           }),
+        ReactiveFormsModule,
         BrowserModule,
         BrowserAnimationsModule,
         AppRoutingModule,
@@ -39,7 +42,9 @@ import { BlockUIModule } from 'ng-block-ui';
         MenuModule
     ],
     providers: [
-        { provide: LocationStrategy, useClass: HashLocationStrategy }
+        { provide: LocationStrategy, useClass: HashLocationStrategy },
+        { provide: LOCALE_ID, useValue: 'pt-BR'},
+        { provide: HTTP_INTERCEPTORS, multi: true, useClass: TokenInterceptor}
     ],
     bootstrap: [AppComponent]
 })
