+++ frontend/src/app/components/topbar/app.topbar.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { AppComponent } from '../../app.component';
 import { Authentication, User } from '@nuvem/angular-base';
+import { AdminComponent } from 'src/app/admin/admin.component';
 
 @Component({
     selector: 'app-topbar',
@@ -8,7 +8,7 @@ import { Authentication, User } from '@nuvem/angular-base';
 })
 export class AppTopbarComponent {
 
-    constructor(public app: AppComponent, private readonly _authentication: Authentication<User>) {
+    constructor(public app: AdminComponent, private readonly _authentication: Authentication<User>) {
     }
 
     get usuario() {
