+++ frontend/src/app/guard/auth.guard.ts
@@ -1,0 +1,26 @@
+import { Injectable } from '@angular/core';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
+import { Observable } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuard implements CanActivate {
+
+  constructor(private router: Router){}
+
+  canActivate(
+    next: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    
+      let token = localStorage.getItem('token');
+
+      if(token) {
+        return true;
+      }
+      this.router.navigate(['./login']);
+      return false;
+
+  }
+  
+}
