+++ frontend/src/app/service/login.service.ts
@@ -1,0 +1,14 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LoginService {
+
+  constructor(private http: HttpClient) { }
+
+  login(credentials){
+    return this.http.post<any>(`api/login`, credentials);
+  }
+}
