+++ frontend/src/app/usuario/listagem-page/listagem-page.component.css
@@ -1,0 +1,1 @@
+td{word-wrap: break-word}
